v2/requirements.txt
# Yamazaki v2 - Requirements
# AutoGen 0.7.5 (latest) with modern Python practices

# AutoGen Core (latest)
autogen-agentchat>=0.7.0
autogen-core>=0.7.0
autogen-ext>=0.7.0

# Configuration Management
pydantic>=2.11.0
pydantic-settings>=2.7.0
python-dotenv>=1.1.0
pyyaml>=6.0.0

# Database (Async)
sqlalchemy>=2.0.36
aiosqlite>=0.20.0       # Async SQLite
asyncpg>=0.30.0         # Async PostgreSQL (optional)
aiomysql>=0.2.0         # Async MySQL (optional)

# HTTP Client
httpx>=0.28.0

# OpenAI & Azure
openai>=1.96.0

# OpenTelemetry (Observability)
opentelemetry-api>=1.35.0
opentelemetry-sdk>=1.35.0

# Utilities
tenacity>=9.1.0
rich>=14.0.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0
pytest-benchmark>=5.1.0

# Optional: Web Tools (for future QA team)
playwright>=1.50.0      # Browser automation
beautifulsoup4>=4.12.0  # HTML parsing
